Update Invoke-CMApplyDriverPackage.ps1
I changed it so it will work on Acer. We want this script to work with "any" computer. We had been using 2.2.3 before and it "just works".

Acer laptop only returns "model" because Acer sucks.

Acer desktop returns SKU so that's usable if you make the package for it.

Either way, it can fall back to an Acer fallback package. I'm not sure why the fallback package has to be OEM-specific now.

Script still throws back error code 0x80090027 on various packages, I still don't know what that means.